requirements.txt
flask
pandas
scipy
cryptography
requests
numpy
python-Levenshtein
